+++ test/run.js
@@ -1,7 +1,14 @@
+var urun = require('urun');
+
 var options = {};
 
 if (process.env.FILTER) {
   options.include = new RegExp(process.env.FILTER + '.*\\.js$');
 }
 
-require('urun')(__dirname, options);
+options.reporter = 'BashTapReporter';
+options.verbose  = process.env.VERBOSE
+  ? Boolean(JSON.parse(process.env.VERBOSE))
+  : true;
+
+urun(__dirname, options);