+++ test/unit/pool-cluster/test-connection-error.js
@@ -1,0 +1,29 @@
+var assert  = require('assert');
+var common  = require('../../common');
+var cluster = common.createPoolCluster({
+  canRetry             : true,
+  removeNodeErrorCount : 5
+});
+var server  = common.createFakeServer();
+var Errors  = require(common.lib + '/protocol/constants/errors');
+
+var connCount  = 0;
+var poolConfig = common.getTestConfig({port: common.fakeServerPort});
+cluster.add('MASTER', poolConfig);
+
+server.listen(common.fakeServerPort, function(err) {
+  assert.ifError(err);
+
+  cluster.getConnection('MASTER', function(err, conn){
+    assert.ok(err);
+    assert.equal(err.code, 'ER_HOST_NOT_PRIVILEGED');
+    assert.equal(err.fatal, true);
+    assert.equal(connCount, 5);
+    server.destroy();
+  });
+});
+
+server.on('connection', function(incomingConnection) {
+  connCount += 1;
+  incomingConnection.deny('You suck.', Errors.ER_HOST_NOT_PRIVILEGED);
+});